requirements.txt
autopep8==1.6.0
black==22.3.0
certifi==2022.5.18.1
cffi==1.14.5
charset-normalizer==2.0.12
click==8.1.3
cryptography==3.4.7
idna==3.3
mypy-extensions==0.4.3
pathspec==0.9.0
platformdirs==2.5.2
protobuf==3.6.0
pycairo==1.21.0
pycodestyle==2.8.0
pycparser==2.20
PyGObject==3.42.1
python-dotenv==0.20.0
requests==2.28.0
six==1.11.0
toml==0.10.2
tomli==2.0.1
typing_extensions==4.2.0
urllib3==1.26.9
